some clean up and all in on sync in rbc
