+++ archipelago/src/test/java/com/hellblazer/archipeligo/ring/SyncRingCommunicationsTest.java
@@ -13,6 +13,7 @@ import io.grpc.inprocess.InProcessServerBuilder;
 import org.junit.jupiter.api.Test;
 
 import java.io.IOException;
+import java.time.Duration;
 import java.util.UUID;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.Executors;
@@ -77,14 +78,19 @@ public class SyncRingCommunicationsTest {
         var router = new RouterImpl(serverMember1, serverBuilder, cacheBuilder, null);
         var commsA = router.create(serverMember1, context.getId(), new ServiceImpl(local1, "A"), "A", ServerImpl::new, TestItClient::new, local1);
 
-        var commsB = router.create(serverMember2, context.getId(), new ServiceImpl(local2, "B"), "B", ServerImpl::new, TestItClient::new, local2);
+        router.create(serverMember2, context.getId(), new ServiceImpl(local2, "B"), "B", ServerImpl::new,
+                      TestItClient::new, local2);
 
-        router.start();
-        var sync = new SyncRingCommunications<Member, TestItService>(context, serverMember1, commsA);
-        var countdown = new CountDownLatch(1);
-        sync.execute((link, round) -> link.ping(Any.getDefaultInstance()), (result, destination) -> countdown.countDown());
-        assertTrue(countdown.await(1, TimeUnit.SECONDS), "Completed: " + countdown.getCount());
-        assertFalse(pinged1.get());
-        assertTrue(pinged2.get());
+        try {
+            router.start();
+            var sync = new SyncRingCommunications<Member, TestItService>(context, serverMember1, commsA);
+            var countdown = new CountDownLatch(1);
+            sync.execute((link, round) -> link.ping(Any.getDefaultInstance()), (result, destination) -> countdown.countDown());
+            assertTrue(countdown.await(1, TimeUnit.SECONDS), "Completed: " + countdown.getCount());
+            assertFalse(pinged1.get());
+            assertTrue(pinged2.get());
+        } finally {
+            router.close(Duration.ofSeconds(1));
+        }
     }
 }
