+++ archipelago/src/test/java/com/hellblazer/archipeligo/ring/SyncRingIteratorTest.java
@@ -76,21 +76,25 @@ public class SyncRingIteratorTest {
         var serverBuilder = InProcessServerBuilder.forName(name);
         var cacheBuilder = ServerConnectionCache.newBuilder().setFactory(to -> InProcessChannelBuilder.forName(name).build());
         var router = new RouterImpl(serverMember1, serverBuilder, cacheBuilder, null);
-        RouterImpl.CommonCommunications<TestItService, TestIt> commsA = router.create(serverMember1, context.getId(), new ServiceImpl(local1, "A"), "A", ServerImpl::new, TestItClient::new, local1);
+        var commsA = router.create(serverMember1, context.getId(), new ServiceImpl(local1, "A"), "A", ServerImpl::new, TestItClient::new, local1);
 
-        RouterImpl.CommonCommunications<TestItService, TestIt> commsB = router.create(serverMember2, context.getId(), new ServiceImpl(local2, "B"), "B", ServerImpl::new, TestItClient::new, local2);
+        var commsB = router.create(serverMember2, context.getId(), new ServiceImpl(local2, "B"), "B", ServerImpl::new, TestItClient::new, local2);
 
-        router.start();
-        var frequency = Duration.ofMillis(1);
-        var scheduler = Executors.newScheduledThreadPool(100, Thread.ofVirtual().factory());
-        var sync = new SyncRingIterator<Member, TestItService>(frequency, context, serverMember1, scheduler, commsA);
-        var countdown = new CountDownLatch(3);
-        sync.iterate(context.getId(), (link, round) -> link.ping(Any.getDefaultInstance()), (round, result, link) -> {
-            countdown.countDown();
-            return true;
-        });
-        assertTrue(countdown.await(1, TimeUnit.SECONDS));
-        assertFalse(pinged1.get());
-        assertTrue(pinged2.get());
+        try {
+            router.start();
+            var frequency = Duration.ofMillis(1);
+            var scheduler = Executors.newScheduledThreadPool(1, Thread.ofVirtual().factory());
+            var sync = new SyncRingIterator<Member, TestItService>(frequency, context, serverMember1, scheduler, commsA);
+            var countdown = new CountDownLatch(3);
+            sync.iterate(context.getId(), (link, round) -> link.ping(Any.getDefaultInstance()), (round, result, link) -> {
+                countdown.countDown();
+                return true;
+            });
+            assertTrue(countdown.await(1, TimeUnit.SECONDS));
+            assertFalse(pinged1.get());
+            assertTrue(pinged2.get());
+        } finally {
+            router.close(Duration.ofSeconds(1));
+        }
     }
 }
