+++ rbc/src/test/resources/logback-test.xml
@@ -21,10 +21,10 @@
     <logger name="com.hellblazer.archipelago" level="INFO" additivity="false">
         <appender-ref ref="STDOUT"/>
     </logger>
-    <logger name="com.hellblazer.archipelago.membership.rbc" level="trace" additivity="false">
+    <logger name="com.hellblazer.rbc" level="debug" additivity="false">
         <appender-ref ref="STDOUT"/>
     </logger>
-    <logger name="com.hellblazer.archipelago.membership.rbc.comms" level="INFO" additivity="false">
+    <logger name="com.hellblazer.rbc.comms" level="INFO" additivity="false">
         <appender-ref ref="STDOUT"/>
     </logger>
     <logger name="com.hellblazer.archipelago.ring" level="INFO" additivity="false">
